+++ test/fixtures/util/pipe_processor.js
@@ -1,0 +1,13 @@
+/**
+ * Mock a simple plugin that takes a file input and just strips any linebreak
+ * or return characters.
+ */
+module.exports = function() {
+  return require('through2').obj(function ( targetFile, encoding, done ) {
+    targetFile.contents = new Buffer(
+      targetFile.contents.toString().replace(/\r?\n|\r/g, '')
+    );
+    this.push(targetFile);
+    done();
+  });
+};