+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # gulp-inception changelog
 
+## 0.1.0
+- Added `pipeThrough` option, to act as a type of middleware processor on file
+streams before merging into the target file.
+
 ## 0.0.1
 - Initial release